refactor(#303): extract all colors to theme
