+++ app/lib/common/models/medication/warning_level.dart
@@ -13,16 +13,6 @@ enum WarningLevel {
   ok
 }
 
-extension WarningLevelColor on WarningLevel {
-  static final _colorMap = {
-    WarningLevel.danger.name: Color(0xFFFFAFAF),
-    WarningLevel.warning.name: Color(0xFFFFEBCC),
-    WarningLevel.ok.name: Color(0xFFCFE8CF),
-  };
-
-  Color get color => WarningLevelColor._colorMap[name]!;
-}
-
 extension WarningLevelIcon on WarningLevel {
   static final _iconMap = {
     WarningLevel.danger.name: Icons.dangerous_rounded,
