+++ app/lib/common/widgets/rounded_card.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import '../module.dart';
 
 class RoundedCard extends StatelessWidget {
   RoundedCard({
@@ -23,8 +23,7 @@ class RoundedCard extends StatelessWidget {
 
     return DecoratedBox(
       decoration: BoxDecoration(
-        // TODO(kolioOtSofia): adjust to theme colors when merged)
-        color: Color(0xFFF9F9F9),
+        color: PharMeTheme.surfaceColor,
         border: Border.all(width: 0.5, color: Colors.black.withOpacity(0.2)),
         borderRadius: borderRadius,
         boxShadow: [
